Adding functionality to import and export customers via button

+++ source/src/BusinessLayer/Customers/Commands/ExportCustomersCommand.cs
@@ -3,32 +3,50 @@ using BusinessLayer.Base.Services;
 using BusinessLayer.Base.Stores;
 using DataLayer.Customers.DTOs;
 using System.Text.Json;
+using System.Xml.Serialization;
 
 namespace BusinessLayer.Customers.Commands
 {
     public class ExportCustomersCommand : BaseAsyncCommand
     {
-        public ExportCustomersCommand(ManagerStore managerStore, IFileDialogue fileDialogue)
+        public ExportCustomersCommand(ManagerStore managerStore, IDialogService dialogService)
         {
             _managerStore = managerStore;
-            _fileDialogue = fileDialogue;
+            _dialogService = dialogService;
         }
 
         public override async Task ExecuteAsync(object? parameter)
         {
-            IEnumerable<CustomerDTO> customers = _managerStore.Customers;
+            IEnumerable<SerializableCustomerDTO> customers;
 
-            string fileName = _fileDialogue.SaveFileDialog("Json files (*.json)|*.json|Xml files (*.xml)|*.xml");
+            try
+            {
+                customers = await _managerStore.GetAllSerializableCustomersAsync();
+            }
+            catch (Exception e)
+            {
+                return;
+            }
+
+            string fileName = _dialogService.SaveFileDialog("Json files (*.json)|*.json|Xml files (*.xml)|*.xml");
+
+            if (string.IsNullOrEmpty(fileName))
+                return;
 
             if (File.Exists(fileName))
                 File.Delete(fileName);
 
             using FileStream stream = File.Open(fileName, FileMode.Create, FileAccess.Write);
 
-            JsonSerializer.Serialize(stream, customers, customers.GetType(), JsonSerializerOptions.Default);
+            if (Path.GetExtension(fileName) == ".json")
+                JsonSerializer.Serialize(stream, customers, customers.GetType(), JsonSerializerOptions.Default);
+            else
+                new XmlSerializer(customers.GetType()).Serialize(stream, customers);
+
+            _dialogService.MessageBoxDialog("Export successfully created!");
         }
 
         private readonly ManagerStore _managerStore;
-        private readonly IFileDialogue _fileDialogue;
+        private readonly IDialogService _dialogService;
     }
 }
